Avoid creating problematic ownerReferences.
It's not legal to have cross-namespaced owners, that is, a
cluster scoped object owned by a namespaced object.

To ensure we preserve the same behaviour upon deletion,
create a finalizer and have it clean up the objects.
(Inspired by #122, which was accidentally lost in a refactor.)

To ensure we are still able to watch for events on those
children, switch to EnqueueRequestsFromMapFunc, which always
maps to HCO.

Signed-off-by: Maya Rashish <mrashish@redhat.com>